Pull request addressing https://github.com/facebook/presto/issues/4567

After profiling and investigating the problematic function was `com.facebook.presto.operator.PagesHashStrategy#positionEqualsPosition`, which was taking up over 90% time of `com.facebook.presto.operator.InMemoryJoinHash#InMemoryJoinHash`. 

This pull request consists of two fixes which one speeds up `positionEqualsPosition` globally but only for queries using `optimize_hash_generation=true`. Second fix further improves `InMemoryJoinHash` for both `optimize_hash_generation=true` and `optimize_hash_generation=false`

In our cluster query:
```
select book_id, customer_id, quantity, genre from transactions join books on (book_id = id) where transaction_year >= 2014 and year_published >= 2014 and publisher not in ('Oxford University Press', 'Cambridge University Press') and price + quantity > 300;
```
has sped up around 10 seconds for distributed joins (from ~2minutes), 20 seconds for broadcast joins with `task_hash_build_concurrency=4` (also from ~2minutes) and over one minute with `task_hash_build_concurrency=1` (from over 3 minutes).

JMH results are following:

Base results (before changes):
```
Benchmark                                      (buildHashEnabled)  Mode  Cnt  Score   Error  Units
BenchmarkHashBuildOperator.benchmarkBuildHash               false  avgt   10  8.623 ± 0.168  ms/op
BenchmarkHashBuildOperator.benchmarkBuildHash                true  avgt   10  4.819 ± 0.117  ms/op
```

With optimized `JoinCompiler`:
```
Benchmark                                      (buildHashEnabled)  Mode  Cnt  Score   Error  Units
BenchmarkHashBuildOperator.benchmarkBuildHash               false  avgt   10  8.417 ± 0.125  ms/op
BenchmarkHashBuildOperator.benchmarkBuildHash                true  avgt   10  3.498 ± 0.066  ms/op
```

Final results (with changes in `InMemoryJoinHash`)
```
Benchmark                                      (buildHashEnabled)  Mode  Cnt  Score   Error  Units
BenchmarkHashBuildOperator.benchmarkBuildHash               false  avgt   10  7.182 ± 0.121  ms/op
BenchmarkHashBuildOperator.benchmarkBuildHash                true  avgt   10  3.158 ± 0.378  ms/op
```