@dain 2016-02-17
What is the type of 'book_id'?  I am asking because I would be surprised if there is a speed up for `BIGINT` types, since the hash is the same as the value.  I would actually expect to see a slight slowdown.  Additionally, for joins on unique values, this would also just be extra cost.

If there is a speedup for `BIGINT` types, there should be a bigger win in fixing the general slowness.