@KBP-TDC 2016-02-17
Yes, book_id is BIGNIT. The schema was derived from here: http://docs.aws.amazon.com/ElasticMapReduce/latest/DeveloperGuide/query-impala-create-tables.html