@pnowojski 2016-02-18
@dain here are the results of more detailed benchmark, with single `VARCHAR`, single `BIGINT` or two `BIGINT` and `VARCHAR` as `hashColumns`. Additionally added comparison for non unique values in build table:

Before changes:
```
Benchmark                                      (buildHashEnabled)  (hashColumns)  (valuesRepetition)  Mode  Cnt    Score    Error  Units
BenchmarkHashBuildOperator.benchmarkBuildHash               false        varchar                   1  avgt    5  180.441 ±  6.239  ms/op
BenchmarkHashBuildOperator.benchmarkBuildHash               false        varchar                   5  avgt    5  137.639 ±  2.005  ms/op
BenchmarkHashBuildOperator.benchmarkBuildHash               false         bigint                   1  avgt    5  115.679 ± 13.836  ms/op
BenchmarkHashBuildOperator.benchmarkBuildHash               false         bigint                   5  avgt    5  115.966 ±  9.372  ms/op
BenchmarkHashBuildOperator.benchmarkBuildHash               false            all                   1  avgt    5  207.165 ±  2.333  ms/op
BenchmarkHashBuildOperator.benchmarkBuildHash               false            all                   5  avgt    5  218.213 ±  3.864  ms/op
BenchmarkHashBuildOperator.benchmarkBuildHash                true        varchar                   1  avgt    5  133.810 ±  6.440  ms/op
BenchmarkHashBuildOperator.benchmarkBuildHash                true        varchar                   5  avgt    5  127.020 ±  5.199  ms/op
BenchmarkHashBuildOperator.benchmarkBuildHash                true         bigint                   1  avgt    5  107.059 ±  2.086  ms/op
BenchmarkHashBuildOperator.benchmarkBuildHash                true         bigint                   5  avgt    5  113.760 ±  3.454  ms/op
BenchmarkHashBuildOperator.benchmarkBuildHash                true            all                   1  avgt    5  133.882 ±  7.277  ms/op
BenchmarkHashBuildOperator.benchmarkBuildHash                true            all                   5  avgt    5  182.045 ±  4.586  ms/op
```

After changes:
```
Benchmark                                      (buildHashEnabled)  (hashColumns)  (valuesRepetition)  Mode  Cnt    Score    Error  Units
BenchmarkHashBuildOperator.benchmarkBuildHash               false        varchar                   1  avgt    5  112.351 ±  3.736  ms/op
BenchmarkHashBuildOperator.benchmarkBuildHash               false        varchar                   5  avgt    5  144.651 ±  3.644  ms/op
BenchmarkHashBuildOperator.benchmarkBuildHash               false         bigint                   1  avgt    5   75.284 ± 23.733  ms/op
BenchmarkHashBuildOperator.benchmarkBuildHash               false         bigint                   5  avgt    5  115.308 ± 23.932  ms/op
BenchmarkHashBuildOperator.benchmarkBuildHash               false            all                   1  avgt    5  138.974 ±  3.405  ms/op
BenchmarkHashBuildOperator.benchmarkBuildHash               false            all                   5  avgt    5  222.746 ±  7.552  ms/op
BenchmarkHashBuildOperator.benchmarkBuildHash                true        varchar                   1  avgt    5   73.931 ± 46.395  ms/op
BenchmarkHashBuildOperator.benchmarkBuildHash                true        varchar                   5  avgt    5  149.307 ±  7.006  ms/op
BenchmarkHashBuildOperator.benchmarkBuildHash                true         bigint                   1  avgt    5   74.140 ± 47.255  ms/op
BenchmarkHashBuildOperator.benchmarkBuildHash                true         bigint                   5  avgt    5  116.149 ±  3.439  ms/op
BenchmarkHashBuildOperator.benchmarkBuildHash                true            all                   1  avgt    5   72.318 ± 18.933  ms/op
BenchmarkHashBuildOperator.benchmarkBuildHash                true            all                   5  avgt    5  181.415 ±  6.184  ms/op
```

Keep in mind that the most significant improvement are not the changes in `JoinCompiler` but added quick look up array in `InMemoryJoinHash`. 

So indeed for joins on single `BIGINT` column we can expect large improvements (33%). Today I have spent some time reading JIT'ed code to make sure that there is no unexpected issue with `PagesHashStrategy` and to me it looks like there is no such problem. Simply using array of integers for storing hashes indexed by position is much quicker compared to accessing values stored in some column within `PagesHashStrategy`. Accessing such value (even if it's single `BIGINT`) requires:

1. get `pageAddress from `addresses` for given `position` (including boundary check)
2. decoding `blockIndex` and `blockPosition` from `pageAddress`
3. jumping to block in list of blocks using `blockIndex` (including boundary check)
4. reading value from block using `blockPosition` (including boundary check)

And you have to do all of that twice (for `possition` and `currentKey`) for single call to `positionEqualsPosition`.

Compare that to shortcut:
```
if (positionToHashes[position] == positionToHashes[currentKey] && 
    positionEqualsPosition(currentKey, position))
```
First of all, `positionToHashes[position]` is still in the registry after computing it's value earlier, so all we have to do is:

1. load `positionToHashes[currentKey]` (including boundary check)
2. compare it with value in registry

which basically translates to 5 assembler instructions.

I'm not sure how would did you mean by fixing it more generally? One way that I had in mind (but didn't want to over engineer) is to create wrapper around PagesHashStategy that have  `positionToHashes`  array. Then this same trick that I have used to speed up `InMemoryJoinHash` would be used automatically in whole code base and this array would be re-used.

Or did you have in mind something more hardcore, like changing how the columns are stored in memory?

Generally speaking this optimization works best for columns with unique values. With non unique values there is still an improvement when `buildHashEnabled=false`.