@dain 2016-02-18
This is a really impressive speedup. 

I'm still curious why this is faster.  Assuming we have a low collision rate in the hash table, when we are doing a lookup for a value in the hash, the first cell checked should be the correct cell, so the hash comparison would always pass, so it is just "extra" work.  Are you testing hash misses?

As for the more "general" solution... I was thinking that maybe positionEqualsPosition or maybe even getting a Type value from a block was overly expensive.

So one possible issue is we are increasing the size of the hash by 4 bytes per entry.  This seems like a fair trade off for such a big perf gain to me, but I'll check if @martint or @cberner think it is a problem.