@pnowojski 2016-02-19
Define what do you mean by "low collision rate"? Currently it's around 40%-50% for joins on column with unique values. In this JMH benchmark with 1 000 000 positions, we now have 455 207 collisions with hash table of size 2 097 152. To reduce collisions, we need bigger hash table. When I rise it's size to 4 194 304, we still have 156 679 collisions for 1 000 000 entries. That's poor trade off considered increased memory footprint by 2097152 * 4 bytes. Additional `positionToHashes` array takes only 1000 000 * 4 bytes and reduce number of calls to `positionEqualsPosition` (by far the most costly effect of a collision) down to 98.