@pnowojski 2016-02-25
I have small update @dain. 

First of all, I have dropped the change in the `JoinCompiler`, it didn't make sense with extensive use of `pagesToHashes` array. 

Secondly I have found out that constructing `positionToHashes` is much faster when it's done is separate loop. I'm not sure why, but it is (benchmarks below). I was looking into JIT'ed code, but hasn't found any clue, code looks basically similar in both versions, so it might have something to do with memory access. Or maybe when there are two loops instead of one, each one of them can use the CPU registers more effectively (with less offloading to memory?). 

As a reminder JMH results of hash build performance with `InMemoryJoinHash's` construction in one loop, but already with `positionToHashes` optimization. (the following table is copied paste from one of my previous posts): 

```
Benchmark                                      (buildHashEnabled)  (hashColumns)  (valuesRepetition)  Mode  Cnt    Score    Error  Units
BenchmarkHashBuildOperator.benchmarkBuildHash               false        varchar                   1  avgt    5  112.400 ± 17.160  ms/op
BenchmarkHashBuildOperator.benchmarkBuildHash               false        varchar                   5  avgt    5  145.207 ±  5.241  ms/op
BenchmarkHashBuildOperator.benchmarkBuildHash               false         bigint                   1  avgt    5   72.081 ±  3.560  ms/op
BenchmarkHashBuildOperator.benchmarkBuildHash               false         bigint                   5  avgt    5  112.367 ±  3.709  ms/op
BenchmarkHashBuildOperator.benchmarkBuildHash               false            all                   1  avgt    5  135.116 ±  2.160  ms/op
BenchmarkHashBuildOperator.benchmarkBuildHash               false            all                   5  avgt    5  206.728 ±  7.362  ms/op
BenchmarkHashBuildOperator.benchmarkBuildHash                true        varchar                   1  avgt    5   67.325 ±  1.794  ms/op
BenchmarkHashBuildOperator.benchmarkBuildHash                true        varchar                   5  avgt    5  139.034 ±  1.684  ms/op
BenchmarkHashBuildOperator.benchmarkBuildHash                true         bigint                   1  avgt    5   65.619 ±  1.637  ms/op
BenchmarkHashBuildOperator.benchmarkBuildHash                true         bigint                   5  avgt    5  100.291 ±  3.145  ms/op
BenchmarkHashBuildOperator.benchmarkBuildHash                true            all                   1  avgt    5   68.186 ±  7.439  ms/op
BenchmarkHashBuildOperator.benchmarkBuildHash                true            all                   5  avgt    5  190.411 ±  6.763  ms/op
```

The same benchmark with two separate loops in the constructor:
```
Benchmark                                              (buildHashEnabled)  (hashColumns)  (valuesRepetition)  Mode  Cnt    Score   Error  Units
BenchmarkHashBuildAndJoinOperators.benchmarkBuildHash               false        varchar                   1  avgt    5   71.624 ± 5.355  ms/op
BenchmarkHashBuildAndJoinOperators.benchmarkBuildHash               false        varchar                   5  avgt    5  128.337 ± 4.694  ms/op
BenchmarkHashBuildAndJoinOperators.benchmarkBuildHash               false         bigint                   1  avgt    5   47.119 ± 1.374  ms/op
BenchmarkHashBuildAndJoinOperators.benchmarkBuildHash               false         bigint                   5  avgt    5  101.466 ± 3.613  ms/op
BenchmarkHashBuildAndJoinOperators.benchmarkBuildHash               false            all                   1  avgt    5   86.438 ± 2.214  ms/op
BenchmarkHashBuildAndJoinOperators.benchmarkBuildHash               false            all                   5  avgt    5  186.427 ± 9.107  ms/op
BenchmarkHashBuildAndJoinOperators.benchmarkBuildHash                true        varchar                   1  avgt    5   52.610 ± 1.964  ms/op
BenchmarkHashBuildAndJoinOperators.benchmarkBuildHash                true        varchar                   5  avgt    5  112.519 ± 2.538  ms/op
BenchmarkHashBuildAndJoinOperators.benchmarkBuildHash                true         bigint                   1  avgt    5   43.787 ± 7.321  ms/op
BenchmarkHashBuildAndJoinOperators.benchmarkBuildHash                true         bigint                   5  avgt    5   95.850 ± 2.471  ms/op
BenchmarkHashBuildAndJoinOperators.benchmarkBuildHash                true            all                   1  avgt    5   52.491 ± 0.706  ms/op
BenchmarkHashBuildAndJoinOperators.benchmarkBuildHash                true            all                   5  avgt    5  152.818 ± 3.023  ms/op
```

Lastly, I have used `positionToHashes` in `InMemoryJoinHash.getJoinPosition` and this improved the actual hash joining by 10%-25%. So summing up, here are the latest results of both hash building and hash joining. In the following results, I have changed the number of pages from 1 000 000 down to 700 000.

Results before my changes:
```
Benchmark                                             (buildHashEnabled)  (hashColumns)  Mode  Cnt    Score    Error  Units
BenchmarkHashBuildOperator.benchmarkBuildAndJoinHash               false        varchar  avgt    5  537.103 ± 30.008  ms/op
BenchmarkHashBuildOperator.benchmarkBuildAndJoinHash               false         bigint  avgt    5  427.539 ± 16.921  ms/op
BenchmarkHashBuildOperator.benchmarkBuildAndJoinHash               false            all  avgt    5  592.554 ± 18.444  ms/op
BenchmarkHashBuildOperator.benchmarkBuildAndJoinHash                true        varchar  avgt    5  526.489 ± 25.400  ms/op
BenchmarkHashBuildOperator.benchmarkBuildAndJoinHash                true         bigint  avgt    5  444.495 ± 16.074  ms/op
BenchmarkHashBuildOperator.benchmarkBuildAndJoinHash                true            all  avgt    5  545.993 ± 29.077  ms/op
```

Optimized HashBuildOperator (InMemoryJoinHash constructor):
```
Benchmark                                             (buildHashEnabled)  (hashColumns)  Mode  Cnt    Score    Error  Units
BenchmarkHashBuildOperator.benchmarkBuildAndJoinHash               false        varchar  avgt    5  434.505 ± 20.431  ms/op
BenchmarkHashBuildOperator.benchmarkBuildAndJoinHash               false         bigint  avgt    5  352.301 ± 13.553  ms/op
BenchmarkHashBuildOperator.benchmarkBuildAndJoinHash               false            all  avgt    5  467.271 ± 13.389  ms/op
BenchmarkHashBuildOperator.benchmarkBuildAndJoinHash                true        varchar  avgt    5  433.748 ± 23.855  ms/op
BenchmarkHashBuildOperator.benchmarkBuildAndJoinHash                true         bigint  avgt    5  378.929 ± 26.891  ms/op
BenchmarkHashBuildOperator.benchmarkBuildAndJoinHash                true            all  avgt    5  450.636 ± 25.388  ms/op
```

All changes together - optimized both HashBuildOperator and HashJoinOperator:
```
Benchmark                                             (buildHashEnabled)  (hashColumns)  Mode  Cnt    Score     Error  Units
BenchmarkHashBuildOperator.benchmarkBuildAndJoinHash               false        varchar  avgt    5  333.764 ±  13.854  ms/op
BenchmarkHashBuildOperator.benchmarkBuildAndJoinHash               false         bigint  avgt    5  288.131 ±  12.473  ms/op
BenchmarkHashBuildOperator.benchmarkBuildAndJoinHash               false            all  avgt    5  369.712 ± 26.822  ms/op
BenchmarkHashBuildOperator.benchmarkBuildAndJoinHash                true        varchar  avgt    5  334.241 ±  13.941  ms/op
BenchmarkHashBuildOperator.benchmarkBuildAndJoinHash                true         bigint  avgt    5  315.357 ±  14.749  ms/op
BenchmarkHashBuildOperator.benchmarkBuildAndJoinHash                true            all  avgt    5  344.928 ±  28.220  ms/op
```